refactor : 관련된 recoil state를 한곳으로 모았다.
